requirements.txt
pygame>=2.5.0
python-rtmidi>=1.5.0
pyserial>=3.5
bleak>=0.22.0
websocket-client>=1.8.0
Pillow>=10.0.0
opencv-python>=4.8.0
